genData.R

Generate competing risks interval-censored data under the proportional odds given a baseline hazard function and some information about observation times.

genData(
  seed = NULL,
  n,
  alpha1,
  alpha2,
  beta1,
  beta2,
  obsTimes = seq(from = 4, to = 28, by = 4),
  probMiss = 0.1,
  windowHalf = 1
)

Arguments

seed

Random seed.

n

The number of observations to generate (sample size).

alpha1

First parameter for Gompertz model.

alpha2

Second parameter for Gompertz model.

beta1

Third parameter for Gompertz model.

beta2

Fourth parameter for Gompertz model.

obsTimes

Vector of the intended observation times.

probMiss

The probability of missing any given visit (default is 0.1).

windowHalf

The amount of time before or after the intended obsTimes that a visit might take place (default is 1).

Value

A list with the elements:

tempTime

n*1 vector of exact event times.

tempType

n*1 vector of exact event types

deltaVecSimple

n*1 vector of whether the event was observed after follow-up started (t>0).

deltaVec

n*1 vector of the event cause (cause 1 or 2; or right-censored 0).

leftTimes

n*1 vector of left side of interval times.

rightTimes

n*1 vector of right side of interval times.

Examples

seed <- 2022
n <- 1000
alpha1 <- -0.058
alpha2 <- -0.035
beta1 <- 0.03
beta2 <- log(1 - exp(beta1 / alpha1)) * alpha2
outcomeDat <- genData(seed=seed, n=n, alpha1=alpha1, alpha2=alpha2, beta1=beta1, beta2=beta2,
obsTimes=seq(from=4, to=28, by=4), probMiss=0.1, windowHalf=1)